Fixes and improvements with tailoring pod and its linkables related to power
Firstly, a few of the linkables that previously didn't need to be powered on need to be to work. If they either didn't have power or were turned off still worked. The linkables that didn't require power were the ones that allowed for extra powers, selection between 2 powers, and removal of weaknesses.

The check for power returns true if there's no `CompPowerTrader` to allow submods to add no power linkables or modify existing ones, mainly for mods aiming for a low tech/magical theme.

Second, `CompGeneTailoringPod` will now require power/fuel if `CompPowerTrader`/`CompRefuelable` are present. Won't change anything as-it-is right now, but will allow submods to give the pod a low tech/magical theme (or at least achieve it easier).

Third, `CompGeneTailoringPod` will now have extra safety around operations being unusable. The gizmo opening a float menu with list of operations will now return a disable option informing of no valid operations (this will prevent errors related to opening float menu with no options). On top of that, `StartOperation` also checks if the operation is valid before proceeding further. This is to fix situations where a machine was turned off while the float menu was still open - in which case the option was still selectable.

The last part could have a slightly different approach - disabling the gizmo itself instead of showing a single disabled option for the float menu. The issue with it is that tooltip showing `disabledReason` for the gizmo expects a label to be present by having 2 empty lines before `disabledReason`. Since this gizmo doesn't have a label, it'll end up with 2 empty lines at the start of the tooltip.