Update supply slingshot trash handling
To sum up the changes:
- Anything with no market value will be sent back
- All trash from VRE will be sent back
- The basic trash from VRE will send 5% extra, but nothing else

Previously, supply slingshot was hardcoded to send back all toxic wastepacks back to the player (and give a 5% extra on top).

With the changes here, the toxic wastepacks are no longer hardcoded. Instead, anything that has a market value of 0 will be automatically sent back at a 1-to-1 ratio.

There's no extra being sent this time due to toxic wastepacks potentially being useful due to the ability to recycle them with Vanilla Recycling Expanded.

On top of that there's also a Dictionary of `ThingDef`s that will be sent back (regardless of market value), with the values of the dictionary being the multiplier that will be sent back to the player (if set to 0 then nothing will be sent back and no value will be gained). The VRE trash will return 5% extra since it's mostly worthless.

Other changes include adding `StaticConstructorOnStartup` (to setup the Dictionary) and the value of 60000 in `TicksToReturn was replaced with `GenDate.TicksPerDay` (to make the intent more clear here).